requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
uuid7==0.1.0
gunicorn==22.0.0
django-ratelimit==4.1.0
dj-database-url==2.2.0
PyJWT==2.9.0
cryptography==43.0.1
